Use global URL instead of absolute server path
The way the paths were being referenced, the app would only work
if it was installed in the root folder. Unfortunately, that's not
always the case, and was not the case when I deployed the app
on a server, where a subfolder was the only option.

This commit replaces references to the HTTP_HOST variable, and
uses a globally-included variable instead.